+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Features
 
 ## Fixes and Improvements
+* Switch cpu backend to llama.cpp: https://github.com/TabbyML/tabby/pull/638
+* add `server.completion_timeout` to control the code completion interface timeout: https://github.com/TabbyML/tabby/pull/637
 
 # v0.4.0